drawer components refactor
